Working on combat and player behavior
A lot's changed here. Animations are now grouped into sets to avoid a
lot of repetitive code, so I renamed some of the anim files. The input
class now checks which of the keys currently being pressed was pressed
most recently, that way you can determine which way to face. Will do the
same for movement soon.