Fixed some pernicious bugs, added some features
knockback now respects direction, but for some reason behaves a little
weird on the enemies.

Player attack also respects direction, but there are some offsets I need
to apply to get it to line up correctly.

Enemies now collided with the tile map.

Also began working on the audio player.